portal-search-tuning-blueprints-content-analysis-microsoft-cognitive-services: Package rename
